modified:   CAREER_COUNSELLING/CAREER_COUNSELLING/__pycache__/settings.cpython-312.pyc 	modified:   CAREER_COUNSELLING/CAREER_COUNSELLING/__pycache__/urls.cpython-312.pyc 	modified:   CAREER_COUNSELLING/CAREER_COUNSELLING/settings.py 	modified:   CAREER_COUNSELLING/CAREER_COUNSELLING/urls.py 	modified:   CAREER_COUNSELLING/LOGIN_AUTH_BOT/__pycache__/urls.cpython-312.pyc 	new file:   CAREER_COUNSELLING/LOGIN_AUTH_BOT/__pycache__/views.cpython-312.pyc 	new file:   CAREER_COUNSELLING/LOGIN_AUTH_BOT/urls.py 	new file:   CAREER_COUNSELLING/LOGIN_AUTH_BOT/views.py 	modified:   CAREER_COUNSELLING/WEB/__pycache__/urls.cpython-312.pyc 	modified:   CAREER_COUNSELLING/WEB/__pycache__/views.cpython-312.pyc 	modified:   CAREER_COUNSELLING/WEB/templates/home.html 	modified:   CAREER_COUNSELLING/WEB/urls.py 	modified:   CAREER_COUNSELLING/WEB/views.py 	modified:   CAREER_COUNSELLING/db.sqlite3 	modified:   virenv/lib/python3.12/site-packages/django/contrib/admin/__pycache__/forms.cpython-312.pyc 	modified:   virenv/lib/python3.12/site-packages/django/contrib/staticfiles/__pycache__/storage.cpython-312.pyc 	modified:   virenv/lib/python3.12/site-packages/django/core/cache/backends/__pycache__/memcached.cpython-312.pyc
