+++ CAREER_COUNSELLING/CAREER_COUNSELLING/urls.py
@@ -20,5 +20,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path('', include('WEB.urls'))
+    path('', include('LOGIN_AUTH_BOT.urls')),
 ]
